// src/app/models/input-error.ts
export class InputError {
    mealNameError: boolean = false;
    mealDesc: boolean = false;
    ingredientNameError: boolean = false;
    ingredientDescError: boolean = false;
    amountError: boolean = false;
    amountTypeError: boolean = false;
    caloriesError: boolean = false;
    proteinError: boolean = false;
    fatError: boolean = false;
    measurementInputError: boolean = false;
    
    mealNameErrorDescription: string = '';
    mealDescErrorDescription: string = '';
    ingredientNameErrorDescription: string = '';
    ingredientDescErrorDescription: string = '';
    amountErrorDescription: string = '';
    amountTypeErrorDescription: string = '';
    caloriesErrorDescription: string = '';
    proteinErrorDescription: string = '';
    fatErrorDescription: string = '';
    measurementInputErrorDescription: string = '';
}